Extract the provider-generic checkpoint plumbing (#256)
Mapillary's tile census needs a checkpoint (#256) and no provider module
may import from another's, so the pieces of #239's machinery that do not
depend on WHAT is being crawled move to their own module: where
checkpoints go, how one directory is named, how a rename into it is made
durable, how a stored bbox is compared, and how a finished or
never-opened directory is removed.

What stays in download_kartaview.py is everything shaped by the crawl --
SweepCheckpoint, load_checkpoint's validation cascade, the commit and
purge helpers, and its own format-version and part-name constants. The
section header keeps its KartaView-specific paragraphs (the layout, the
parquet-not-CSV argument, the commit-record-is-a-count rule) and points
at the new home for the three that were never KartaView's: date-free
paths, the channel in the path key, and caller-discards-after-durable.

Those three are re-homed rather than copied verbatim -- they were
written when KartaView was the only checkpointing provider, so the
channel paragraph now names both channel PAIRS and discard_checkpoint's
warning no longer says "KartaView". _state_path comes along too: it is
one line over CHECKPOINT_STATE_FILENAME and both providers need it.

Every moved name is re-exported from download_kartaview in the `X as X`
form, so cli.py, tests/conftest.py and every kv.*/dk.* reference resolve
unchanged.

Pure refactor: 1,609 tests pass with ZERO test edits, which is the
acceptance condition for this step.
